Instantaneous Sync When Joining
Now when a new socket connects to the room it will sync directly to the
current video and time. This is done by assigning host status to the
first socket to join a room. The new socket retrieves the needed
information from this socket and jumps to the correct spot!