Autoscroll, light fix, front end (#74, #75)
A few things were fixed in this commit.

Chat now autoscrolls with each message sent. The turn off the light
button now works properly and will work when clicking on the icon. Small
front end improvements were made including the removal of scroll bars
and a small border on the chat area.