App Stability Update (#43)
The issue that caused the crashes were the reading of room objects that
did not exist. This would only occur on a sleeping heroku server. This
means that when a user joined a room, the server went to sleep, then the
user attempted to interact with the now nonexistent room. This caused a
crash. Just added conditions to check the room before performing any
actions.